requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
Authlib==1.2.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.5
comm==0.1.3
cryptography==41.0.2
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.101.1
fastjsonschema==2.17.1
fqdn==1.5.1
frozenlist==1.4.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
ipykernel==6.24.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.4
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.1
multidict==6.0.4
mutagen==1.46.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.1
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.3
openai==0.27.8
orjson==3.9.5
overrides==7.3.1
packaging==23.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.8.1
pluggy==1.2.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pycryptodomex==3.18.0
pydantic==2.0.3
pydantic-extra-types==2.0.0
pydantic-settings==2.0.3
pydantic_core==2.3.0
pydub==0.25.1
Pygments==2.15.1
pypdf==3.12.2
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.29.1
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.8.10
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
starlette==0.27.0
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.7.1
ujson==5.8.0
uri-template==1.3.0
urllib3==2.0.3
uvicorn==0.23.0
uvloop==0.17.0
validators==0.20.0
watchfiles==0.19.0
wcwidth==0.2.6
weaviate-client==3.22.1
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
widgetsnbextension==4.0.8
Wikipedia-API==0.6.0
yarl==1.9.2
yt-dlp==2023.7.6
